some usability updates
adding initial support for specifying path/to/package for buildit
and cleanit. trying just cd'ing to the directory if necessary and cd'ing
back afterwards and shifting the arg list.

adding some things to addit that seemed more useful, like supporting
a -doc extension and updating the dep file, along with updates to
how the .info file is generated.