remove download stamp if version changed
Why didn't it already do this!?!?!
I hope this way of handling it makes sense, but it is really annoying to have to manually delete
the download stamp after a version change. This might possibly delete the stamp
at a wrong time in some situations, but if you're running cleanit in the middle of the
process, is that really so bad?